requirements.txt
# Web Framework
Flask==2.3.3
Werkzeug==2.3.7

# HTTP Requests
requests==2.31.0

# Image Processing
Pillow==10.0.1

# Barcode Detection
pyzbar==0.1.9

# Computer Vision (for camera scanning)
opencv-python==4.8.1.78
numpy==1.24.3

# AI/LLM Integration
langchain-google-genai==1.0.8
langchain-core==0.2.38
pydantic==2.5.3

# Environment Variables (optional but recommended)
python-dotenv==1.0.0